requirements.txt
alabaster==0.7.10
asn1crypto==0.24.0
astroid==1.6.3
Babel==2.5.3
backports.functools-lru-cache==1.5
bcrypt==3.1.4
boto3==1.7.6
botocore==1.10.6
caniusepython3==6.0.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
configobj==5.0.6
configparser==3.5.0
coverage==4.5.1
cryptography==2.2.2
distlib==0.2.7
dj-database-url==0.5.0
Django==2.0.4
django-appconf==1.0.2
django-avatar==4.1.0
django-classy-tags==0.8.0
django-collectstatic-bower==0.1.2
django-forms-bootstrap==3.1.0
django-storages==1.6.6
django-tinymce==2.7.0
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
futures==3.1.1
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
imagesize==1.0.0
isort==4.3.4
Jinja2==2.10
jmespath==0.9.3
keyring==12.0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.12.2
MarkupSafe==1.0
mccabe==0.6.1
oauth==1.0.1
olefile==0.45.1
packaging==17.1
paramiko==2.4.1
pbr==4.0.2
Pillow==5.1.0
pinax-testimonials==2.0.4
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pyasn1==0.4.2
pycairo==1.17.0
pycparser==2.18
pycryptodome==3.6.4
Pygments==2.2.0
pylint==1.8.4
PyNaCl==1.2.1
pyparsing==2.2.0
pytest-runner==4.2
python-dateutil==2.7.2
python-http-client==3.1.0
python-memcached==1.59
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
SecretStorage==2.3.1
simplejson==3.14.0
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
stripe==1.79.1
testresources==2.0.1
tox==3.0.0
typing==3.6.4
urllib3==1.22
virtualenv==15.2.0
wadllib==1.3.2
webencodings==0.5.1
whitenoise==3.3.1
wrapt==1.10.11
zope.interface==4.5.0
